Moves page components to their own file structure adds some more optional chaining
