+++ app/javascript/packs/MainApp/pages/PrivacyPolicy.jsx
